SFA: Split texture handling into its own file
